OrderManagementTemplateコンポーネントの作成とpage.tsxのリファクタリング
- OrderManagementTemplate.tsxを作成し、注文管理の主要なロジックとUIを移動
- page.tsxをシンプル化し、OrderManagementTemplateを読み込む構成に変更
- コードの可読性とメンテナンス性を向上
